pyproject.toml
[project]
name = "mkdocs-dataview"
authors = [
  { name="Alexander Nevskiy", email="kepkin@gmail.com" },
]
description = "A mkdocs plugin inspired by dataview plugin in Obsidian"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

dynamic = ["version"]

dependencies = [
  "lark == 1.2.2",
  "python-frontmatter == 1.1.0",
  "mkdocs == 1.6.1",
  "markdown-callouts",
  "mdx-gh-links",
  "mkdocs",
  "mkdocs-click",
  "pymdown-extensions",
]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project.entry-points."mkdocs.plugins"]
dataview = "mkdocs_dataview.plugin:DataViewPlugin"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.setuptools.dynamic]
version = {attr = "mkdocs_dataview.__version__"}
